backend/requirements.txt
flask==3.0.2
python-dotenv==1.0.1
langchain==0.1.12
langchain-core==0.1.32
langchain-openai==0.0.8
openai==1.14.2
flask-cors==4.0.0
waitress==3.0.0
httpx==0.27.0
nest-asyncio==1.6.0